Move timer into next button to prevent re-renders
